some more refactoring :D
